+++ internal/provider/helper_test.go
@@ -12,8 +12,7 @@ import (
 	"github.com/slok/terraform-provider-onepasswordorg/internal/model"
 )
 
-// getFakeRepoTmpFile returns a temp file that can be used for the fake
-// repository storage.
+// getFakeRepoTmpFile returns a temp file that can be used for the fake repository storage.
 // It returns the file path and a delete file function.
 func getFakeRepoTmpFile(prefix string) (path string, delete func()) {
 	// Prepare fake storage.
@@ -24,16 +23,12 @@ func getFakeRepoTmpFile(prefix string) (path string, delete func()) {
 	return f.Name(), func() { _ = os.Remove(f.Name()) }
 }
 
-// assertUserOnFakeStorage is a helper to assert the expected user is stored on the fake
-// repository.
 func assertUserOnFakeStorage(t *testing.T, expUser *model.User) resource.TestCheckFunc {
 	assert := assert.New(t)
 
 	return resource.TestCheckFunc(func(s *terraform.State) error {
-		// Get fake repo.
 		repo := getFakeRepository(t)
 
-		// Check user.
 		gotUser, err := repo.GetUserByID(context.TODO(), expUser.ID)
 		assert.NoError(err)
 		assert.Equal(expUser, gotUser)
@@ -41,32 +36,24 @@ func assertUserOnFakeStorage(t *testing.T, expUser *model.User) resource.TestChe
 	})
 }
 
-// assertUserDeletedOnFakeStorage is a helper to assert the expected user ID is not stored on
-// the fake repository.
 func assertUserDeletedOnFakeStorage(t *testing.T, userID string) resource.TestCheckFunc {
 	assert := assert.New(t)
 
 	return resource.TestCheckFunc(func(s *terraform.State) error {
-		// Get fake repo.
 		repo := getFakeRepository(t)
 
-		// Check user is missing.
 		_, err := repo.GetUserByID(context.TODO(), userID)
 		assert.Error(err)
 		return nil
 	})
 }
 
-// assertGroupOnFakeStorage is a helper to assert the expected group is stored on the fake
-// repository.
 func assertGroupOnFakeStorage(t *testing.T, expGroup *model.Group) resource.TestCheckFunc {
 	assert := assert.New(t)
 
 	return resource.TestCheckFunc(func(s *terraform.State) error {
-		// Get fake repo.
 		repo := getFakeRepository(t)
 
-		// Check group.
 		gotGroup, err := repo.GetGroupByID(context.TODO(), expGroup.ID)
 		assert.NoError(err)
 		assert.Equal(expGroup, gotGroup)
@@ -74,32 +61,24 @@ func assertGroupOnFakeStorage(t *testing.T, expGroup *model.Group) resource.Test
 	})
 }
 
-// assertGroupDeletedOnFakeStorage is a helper to assert the expected group ID is not stored on
-// the fake repository.
 func assertGroupDeletedOnFakeStorage(t *testing.T, groupID string) resource.TestCheckFunc {
 	assert := assert.New(t)
 
 	return resource.TestCheckFunc(func(s *terraform.State) error {
-		// Get fake repo.
 		repo := getFakeRepository(t)
 
-		// Check group is missing.
 		_, err := repo.GetGroupByID(context.TODO(), groupID)
 		assert.Error(err)
 		return nil
 	})
 }
 
-// assertGroupMemberOnFakeStorage is a helper to assert the expected membership is stored on the fake
-// repository.
 func assertGroupMemberOnFakeStorage(t *testing.T, expMembership *model.Membership) resource.TestCheckFunc {
 	assert := assert.New(t)
 
 	return resource.TestCheckFunc(func(s *terraform.State) error {
-		// Get fake repo.
 		repo := getFakeRepository(t)
 
-		// Check membership.
 		gotMembership, err := repo.GetMembershipByID(context.TODO(), expMembership.GroupID, expMembership.UserID)
 		assert.NoError(err)
 		assert.Equal(expMembership, gotMembership)
@@ -107,32 +86,24 @@ func assertGroupMemberOnFakeStorage(t *testing.T, expMembership *model.Membershi
 	})
 }
 
-// assertGroupMemberDeletedOnFakeStorage is a helper to assert the expected membership is not stored on
-// the fake repository.
 func assertGroupMemberDeletedOnFakeStorage(t *testing.T, groupID, userID string) resource.TestCheckFunc {
 	assert := assert.New(t)
 
 	return resource.TestCheckFunc(func(s *terraform.State) error {
-		// Get fake repo.
 		repo := getFakeRepository(t)
 
-		// Check group is missing.
 		_, err := repo.GetMembershipByID(context.TODO(), groupID, userID)
 		assert.Error(err)
 		return nil
 	})
 }
 
-// assertVaultOnFakeStorage is a helper to assert the expected vault is stored on the fake
-// repository.
 func assertVaultOnFakeStorage(t *testing.T, expVault *model.Vault) resource.TestCheckFunc {
 	assert := assert.New(t)
 
 	return resource.TestCheckFunc(func(s *terraform.State) error {
-		// Get fake repo.
 		repo := getFakeRepository(t)
 
-		// Check vault.
 		gotVault, err := repo.GetVaultByID(context.TODO(), expVault.ID)
 		assert.NoError(err)
 		assert.Equal(expVault, gotVault)
@@ -140,17 +111,38 @@ func assertVaultOnFakeStorage(t *testing.T, expVault *model.Vault) resource.Test
 	})
 }
 
-// assertVaultDeletedOnFakeStorage is a helper to assert the expected vault ID is not stored on
-// the fake repository.
 func assertVaultDeletedOnFakeStorage(t *testing.T, vaultID string) resource.TestCheckFunc {
 	assert := assert.New(t)
 
 	return resource.TestCheckFunc(func(s *terraform.State) error {
-		// Get fake repo.
 		repo := getFakeRepository(t)
 
-		// Check vault is missing.
 		_, err := repo.GetVaultByID(context.TODO(), vaultID)
+		assert.Error(err)
+		return nil
+	})
+}
+
+func assertVaultGroupAccessOnFakeStorage(t *testing.T, exp *model.VaultGroupAccess) resource.TestCheckFunc {
+	assert := assert.New(t)
+
+	return resource.TestCheckFunc(func(s *terraform.State) error {
+		repo := getFakeRepository(t)
+
+		got, err := repo.GetVaultGroupAccessByID(context.TODO(), exp.VaultID, exp.GroupID)
+		assert.NoError(err)
+		assert.Equal(exp, got)
+		return nil
+	})
+}
+
+func assertVaultGroupAccessDeletedOnFakeStorage(t *testing.T, vaultID, GroupID string) resource.TestCheckFunc {
+	assert := assert.New(t)
+
+	return resource.TestCheckFunc(func(s *terraform.State) error {
+		repo := getFakeRepository(t)
+
+		_, err := repo.GetVaultGroupAccessByID(context.TODO(), vaultID, GroupID)
 		assert.Error(err)
 		return nil
 	})